Complete users

+++ app/Http/Controllers/UserController.php
@@ -39,21 +39,61 @@ class UserController extends Controller
     }
 
     public function update(Request $request, $id){
+        $this->validate($request,[
+            "Avatar"=>"required,image|mimes:jpeg,png,jpg,gif|max:2048"
+        ],[
+            "Avatar.required"=>"Please chose a image",
+            "Avatar.image"=>"Wrong image format",
+            "Avatar.max"=>"Image has max size"
+        ]);
         $data = $request->all();
-        $User = $this->UserRepository->update($data,$id);
-        if($User == true){
-            return redirect()->back()->with('thong_bao','Update an item success!');
+        if($file = $request->hasFile("Avatar")){
+            $file = $request->file("Avatar");
+            $fileExtendtion = $file->getClientOriginalExtension();
+            if($fileExtendtion == "jpg" || $fileExtendtion == "jpeg" || $fileExtendtion == "png" || $fileExtendtion == "gif" || $fileExtendtion == "JPG" || $fileExtendtion == "PNG"){
+                /*
+                 * Kiểm tra và xóa hình ảnh cũ trong thư mục Avatar
+                 * */
+                $checkImage = $this->UserRepository->deleteImage($id);
+                if($checkImage == 0){
+                    return redirect()->back()->with('thong_bao','Update an item failed, please again image!');
+                }else if($checkImage == 2){
+                    return redirect()->back()->with('thong_bao','Update an item failed, image has not exits!');
+                }else if($checkImage == 1){
+                    // Tiến hành gửi hình ảnh vào thư mục
+                    $fileName = str_random(3) . "_" . $file->getClientOriginalName();
+                    if($file->move("upload/Avatar",$fileName)){
+                        $User = $this->UserRepository->update($data,$id);
+                        if($User == true){
+                            return redirect()->back()->with('thong_bao','Update an item success!');
+                        }else{
+                            return redirect()->back()->with('thong_bao','Update an item failed!');
+                        }
+                    }else{
+                        return redirect()->back()->with('thong_bao','Upload image failed, please check again system!');
+                    }
+                }
+            }else{
+                return redirect()->back()->with('thong_bao','Wrong image type!');
+            }
         }else{
-            return redirect()->back()->with('thong_bao','Update an item failed!');
+            return redirect()->back()->with('thong_bao','Please chose a image');
         }
     }
 
     public function destroy($id){
-        $User = $this->UserRepository->delete($id);
-        if($User == true){
-            return redirect('admin/User/Users')->with('thong_bao','Delete an item success!');
-        }else{
-            return redirect('admin/User/Users')->with('thong_bao','Delete an item failed');
+        $checkImage = $this->UserRepository->deleteImage($id);
+        if($checkImage == 1){
+            $User = $this->UserRepository->delete($id);
+            if($User == true){
+                return redirect('admin/User/Users')->with('thong_bao','Delete an item success!');
+            }else{
+                return redirect('admin/User/Users')->with('thong_bao','Delete an item failed');
+            }
+        }else if($checkImage == 0){
+            return redirect('admin/User/Users')->with('thong_bao','Delete an item failed, please check again system');
+        }else if($checkImage == 2){
+            return redirect('admin/User/Users')->with('thong_bao','Delete an item failed, please check again file in system');
         }
     }
